+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/utils/data.py
@@ -730,6 +730,26 @@ class Data:
             tags=["tailscale_extras"],
         ),
         Env(
+            "_ADSBIM_STATE_NETBIRD_LOGIN_LINK",
+            tags=["netbird_ll"],
+            default="",
+        ),
+        Env(
+            "_ADSBIM_STATE_NETBIRD_NAME",
+            tags=["netbird_name"],
+            default="",
+        ),
+        Env(
+            "_ADSBIM_STATE_NETBIRD_EXTRA_ARGS",
+            tags=["netbird_extras"],
+            default="",
+        ),
+        Env(
+            "_ADSBIM_STATE_NETBIRD_SETUP_KEY",
+            tags=["netbird_setup_key", "key"],
+            default="",
+        ),
+        Env(
             "_ADSBIM_STATE_EXTRA_ENV",
             tags=["ultrafeeder_extra_env"],
         ),